[TASK] Move admin check method to Service class
